refactor: add jwtprovider & handle auth exceptions
